requirement.txt
torch
pytorch_lightning
lmdb
tqdm
pillow
timm
scikit-learn